+++ docs/javascripts/mathjax.js
@@ -1,0 +1,16 @@
+window.MathJax = {
+  tex: {
+    inlineMath: [["\\(", "\\)"]],
+    displayMath: [["\\[", "\\]"]],
+  },
+  options: {
+    ignoreHtmlClass: ".*",
+    processHtmlClass: "arithmatex",
+  },
+};
+
+if (typeof document$ !== "undefined") {
+  document$.subscribe(() => {
+    MathJax.typesetPromise();
+  });
+}